setup embalm service
