+++ apps/eaas-api/src/app/services/embalm.service.ts
@@ -1,8 +1,13 @@
 import { ArchaeologistExceptionCode, NodeSarcoClient } from "@sarcophagus-org/sarcophagus-v2-sdk";
-import { uploadEncryptedPayloadToArweave } from "./arweave";
-import { ApiBufferType, PreparedEncryptedPayload } from "../../../common/types";
+import { ApiBufferType, PreparedEncryptedPayload } from "../../../../common/types";
 
-export interface EmbalmOptions {
+interface ArweaveUploadArgs {
+  sarco: NodeSarcoClient;
+  archaeologistPublicKeys: string[];
+  preparedEncryptedPayload: PreparedEncryptedPayload;
+}
+
+interface EmbalmOptions {
   chainId: number;
   sarcophagusName: string;
   resurrectionTime: number;
@@ -10,9 +15,44 @@ export interface EmbalmOptions {
   preparedEncryptedPayload: PreparedEncryptedPayload;
 }
 
-export function formatPreparedEncryptedPayload(
-  arg: PreparedEncryptedPayload,
-): PreparedEncryptedPayload {
+const uploadEncryptedPayloadToArweave = async (args: ArweaveUploadArgs) => {
+  const { sarco, archaeologistPublicKeys } = args;
+
+  const {
+    recipientPublicKey,
+    encryptedPayload,
+    innerEncryptedkeyShares,
+    encryptedPayloadMetadata,
+  } = args.preparedEncryptedPayload;
+
+  return new Promise<string>(async (resolve, reject) => {
+    try {
+      const uploadPromise = sarco.api.uploadPreEncryptedPayloadToArweave({
+        archaeologistPublicKeys,
+        onStep: (step: string) => console.log(`Uploading To Arweave: ${step}`),
+        recipientPublicKey,
+        onUploadChunk: (_: any, chunkedUploadProgress: number) => {
+          console.log(`Upload Progress: ${chunkedUploadProgress}%`);
+        },
+        onUploadChunkError: (msg: string) => {
+          console.error(msg);
+          throw new Error(msg);
+        },
+        onUploadComplete: (uploadId: string) => resolve(uploadId),
+        encryptedPayload: encryptedPayload as Buffer,
+        innerEncryptedkeyShares: innerEncryptedkeyShares as Buffer[],
+        encryptedPayloadMetadata,
+      });
+
+      await uploadPromise;
+    } catch (error: any) {
+      console.log(error);
+      throw new Error(error.message || "Error uploading payload to Bundlr");
+    }
+  });
+};
+
+function formatPreparedEncryptedPayload(arg: PreparedEncryptedPayload): PreparedEncryptedPayload {
   return {
     encryptedPayload: (arg.encryptedPayload as ApiBufferType).data,
     innerEncryptedkeyShares: (arg.innerEncryptedkeyShares as ApiBufferType[]).map(
@@ -23,7 +63,7 @@ export function formatPreparedEncryptedPayload(
   };
 }
 
-export async function runEmbalm(options: EmbalmOptions) {
+async function runEmbalm(options: EmbalmOptions) {
   const {
     chainId,
     sarcophagusName,
@@ -123,3 +163,8 @@ export async function runEmbalm(options: EmbalmOptions) {
     throw new Error(e);
   }
 }
+
+export const embalmService = {
+  formatPreparedEncryptedPayload,
+  runEmbalm,
+};
