+++ apps/eaas-api/src/app/services/index.ts
@@ -2,3 +2,4 @@ export { authService } from "./auth.service";
 export { invitationService } from "./invitation.service";
 export { tokenService } from "./token.service";
 export { userService } from "./user.service";
+export { embalmService } from "./embalm.service";
